ml/requirements.txt
torch>=2.3.0
torchvision>=0.18.0
numpy>=1.26.0
pillow>=10.0.0
tqdm>=4.66.0
matplotlib>=3.8.0
opencv-python>=4.9.0.0
requests>=2.31.0
